Move FloatValidator
